Cria backend + frontend do menu despesas

+++ backend/mifortuna-back/src/despesas/despesas.service.ts
@@ -1,10 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { MoreThanOrEqual, Repository } from 'typeorm';
 import { Despesa } from './entities/despesa.entity';
 import { CreateDespesaDto } from './dto/create-despesa.dto';
 import { UpdateDespesaDto } from './dto/update-despesa.dto';
 import { ILike } from 'typeorm';
+import { v4 } from 'uuid';
 
 @Injectable()
 export class DespesasService {
@@ -14,16 +15,39 @@ export class DespesasService {
     ) {}
 
   create(createDespesaDto: CreateDespesaDto, usuarioId) {
-    const despesa = this.despesaRepository.create({
-      ...createDespesaDto,
-      // if(quantidadeParcelas > 1){
-      //   for(let index 0 : index < quantidadeParcelas.length; index++)
-      // }
-      categoria: { id: createDespesaDto.categoria },
-      formaPagamento: { id: createDespesaDto.formaPagamento },
-      usuario: { id: usuarioId },
-    })  
-    return this.despesaRepository.save(despesa);
+    const parcelas = createDespesaDto.quantidadeParcelas
+
+    if (parcelas && parcelas > 1) {
+      const despesasCriadas: Despesa[] = []
+      const grupo = v4();
+
+      for (let i = 0; i < parcelas; i++) {
+        const novaData = new Date(createDespesaDto.dataVencimento)
+        novaData.setMonth(novaData.getMonth() + i)
+
+        const despesa = this.despesaRepository.create({
+          ...createDespesaDto,
+          dataVencimento: novaData,
+          parcelaAtual: i + 1,
+          categoria: { id: createDespesaDto.categoria },
+          formaPagamento: { id: createDespesaDto.formaPagamento },
+          usuario: { id: usuarioId },
+          grupoParcelas: grupo
+
+        })
+        despesasCriadas.push(despesa)
+      }
+
+      return this.despesaRepository.save(despesasCriadas)
+    } else {
+      const despesa = this.despesaRepository.create({
+        ...createDespesaDto,
+        categoria: { id: createDespesaDto.categoria },
+        formaPagamento: { id: createDespesaDto.formaPagamento },
+        usuario: { id: usuarioId },
+      })
+      return this.despesaRepository.save(despesa)
+    }
   }
 
   findAll(usuarioId: number, nome?: string, categoria?: number, formaPagamento?: number, status?: string) {
@@ -44,6 +68,12 @@ export class DespesasService {
     return this.despesaRepository.find({ where, relations: ['categoria','formaPagamento'] });
   }
 
+  findOne(id: number) {
+    const where: any = { usuario: { id: id }};
+
+    return this.despesaRepository.find({ where });
+  }
+
   update(id: number, updateDespesaDto: UpdateDespesaDto) {
     const dados: any = { ...updateDespesaDto };
     if (updateDespesaDto.categoria) {
@@ -55,8 +85,27 @@ export class DespesasService {
     return this.despesaRepository.update(id, dados);
   }
 
-  remove(id: number) {
-    return this.despesaRepository.delete(id)
+  async remove(id: number ) {
+    const despesa = await this.despesaRepository.findOne({ where: { id } });
+
+    if(!despesa){
+      throw new Error('Despesas não encontrada')
+    }
+
+    if (despesa.quantidadeParcelas === null ){
+      return this.despesaRepository.delete(id)
+    }
+    if(despesa.parcelaAtual < despesa.quantidadeParcelas){
+
+      return this.despesaRepository.delete(
+          { 
+            grupoParcelas: despesa.grupoParcelas,
+            parcelaAtual : MoreThanOrEqual( despesa.parcelaAtual)
+          }
+        )
+    }else{
+      return this.despesaRepository.delete(id)
+    }
   }
   
 }
